Fixed more eres' and made the quickSort module
