+++ homework3/hw3.3/frequencySearch.h
@@ -1,4 +1,4 @@
 #pragma once
 
-int frequencySearch(int *searchingArray, int searchingArrayLength);
+int frequencySearch(int *sourceArray, int arrayLength);
 void quickSort(int *sourceArray, int leftBorder, int rightBorder);
