+++ homework3/hw3.3/hw3.3.vcxproj.filters
@@ -21,6 +21,9 @@
     <ClCompile Include="frequencySearch.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="quickSort.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="frequencySearch.h">
